PaddleFSL/examples/optim/requirements.txt
paddlepaddle
loguru
tqdm
numpy
typed-argument-parser
paddlenlp